requirements.txt
flask
Flask-SocketIO
pillow